+++ lib/kitchen/driver/docker.rb
@@ -20,16 +20,19 @@ require 'securerandom'
 require 'uri'
 require 'net/ssh'
 require 'tempfile'
-require File.join(File.dirname(__FILE__), 'docker', 'erb')
+require 'shellwords'
+
+require 'kitchen/driver/base'
+
+require_relative './docker/erb'
 
 module Kitchen
-
   module Driver
-
     # Docker driver for Kitchen.
     #
     # @author Sean Porter <portertech@gmail.com>
-    class Docker < Kitchen::Driver::SSHBase
+    class Docker < Kitchen::Driver::Base
+      include ShellOut
 
       default_config :binary,        'docker'
       default_config :socket,        ENV['DOCKER_HOST'] || 'unix:///var/run/docker.sock'
@@ -42,7 +45,6 @@ module Kitchen
       default_config :run_command,   '/usr/sbin/sshd -D -o UseDNS=no -o UsePAM=no -o PasswordAuthentication=yes ' +
                                      '-o UsePrivilegeSeparation=no -o PidFile=/tmp/sshd.pid'
       default_config :username,      'kitchen'
-      default_config :password,      'kitchen'
       default_config :tls,           false
       default_config :tls_verify,    false
       default_config :tls_cacert,    nil
@@ -52,6 +54,8 @@ module Kitchen
       default_config :wait_for_sshd, true
       default_config :private_key,   File.join(Dir.pwd, '.kitchen', 'docker_id_rsa')
       default_config :public_key,    File.join(Dir.pwd, '.kitchen', 'docker_id_rsa.pub')
+      default_config :build_options, nil
+      default_config :run_options,   nil
 
       default_config :use_sudo do |driver|
         !driver.remote_socket?
@@ -71,10 +75,20 @@ module Kitchen
         !driver.remote_socket?
       end
 
+      default_config :instance_name do |driver|
+        # Borrowed from kitchen-rackspace
+        [
+          driver.instance.name.gsub(/\W/, ''),
+          (Etc.getlogin || 'nologin').gsub(/\W/, ''),
+          Socket.gethostname.gsub(/\W/, '')[0..20],
+          Array.new(8) { rand(36).to_s(36) }.join
+        ].join('-')
+      end
+
       MUTEX_FOR_SSH_KEYS = Mutex.new
 
       def verify_dependencies
-        run_command("#{config[:binary]} >> #{dev_null} 2>&1", :quiet => true)
+        run_command("#{config[:binary]} >> #{dev_null} 2>&1", quiet: true, use_sudo: config[:use_sudo])
         rescue
           raise UserError,
           'You must first install the Docker CLI tool http://www.docker.io/gettingstarted/'
@@ -103,12 +117,17 @@ module Kitchen
 
       def create(state)
         generate_keys
+        state[:username] = config[:username]
         state[:ssh_key] = config[:private_key]
         state[:image_id] = build_image(state) unless state[:image_id]
         state[:container_id] = run_container(state) unless state[:container_id]
         state[:hostname] = remote_socket? ? socket_uri.host : 'localhost'
         state[:port] = container_ssh_port(state)
-        wait_for_sshd(state[:hostname], nil, :port => state[:port]) if config[:wait_for_sshd]
+        if config[:wait_for_sshd]
+          instance.transport.connection(state) do |conn|
+            conn.wait_until_ready
+          end
+        end
       end
 
       def destroy(state)
@@ -136,7 +155,11 @@ module Kitchen
         docker << " --tlscacert=#{config[:tls_cacert]}" if config[:tls_cacert]
         docker << " --tlscert=#{config[:tls_cert]}" if config[:tls_cert]
         docker << " --tlskey=#{config[:tls_key]}" if config[:tls_key]
-        run_command("#{docker} #{cmd}", options.merge(:quiet => !logger.debug?))
+        run_command("#{docker} #{cmd}", options.merge({
+          quiet: !logger.debug?,
+          use_sudo: config[:use_sudo],
+          log_subject: Thor::Util.snake_case(self.class.to_s),
+        }))
       end
 
       def generate_keys
@@ -184,6 +207,7 @@ module Kitchen
           eos
           packages = <<-eos
             ENV DEBIAN_FRONTEND noninteractive
+            ENV container docker
             RUN apt-get update
             RUN apt-get install -y sudo openssh-server curl lsb-release
           eos
@@ -193,29 +217,33 @@ module Kitchen
             ENV container docker
             RUN yum clean all
             RUN yum install -y sudo openssh-server openssh-clients which curl
-            RUN ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key -N ''
-            RUN ssh-keygen -t dsa -f /etc/ssh/ssh_host_dsa_key -N ''
+            RUN [ -f "/etc/ssh/ssh_host_rsa_key" ] || ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key -N ''
+            RUN [ -f "/etc/ssh/ssh_host_dsa_key" ] || ssh-keygen -t dsa -f /etc/ssh/ssh_host_dsa_key -N ''
           eos
         when 'arch'
+          # See https://bugs.archlinux.org/task/47052 for why we
+          # blank out limits.conf.
           <<-eos
-            RUN pacman -Syu --noconfirm
-            RUN pacman -S --noconfirm openssh sudo curl
-            RUN ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
-            RUN ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
+            RUN pacman --noconfirm -Sy archlinux-keyring
+            RUN pacman-db-upgrade
+            RUN pacman --noconfirm -Sy openssl openssh sudo curl
+            RUN [ -f "/etc/ssh/ssh_host_rsa_key" ] || ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
+            RUN [ -f "/etc/ssh/ssh_host_dsa_key" ] || ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
+            RUN echo >/etc/security/limits.conf
           eos
         when 'gentoo'
           <<-eos
-            RUN emerge sync
+            RUN emerge --sync
             RUN emerge net-misc/openssh app-admin/sudo
-            RUN ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
-            RUN ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
+            RUN [ -f "/etc/ssh/ssh_host_rsa_key" ] || ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
+            RUN [ -f "/etc/ssh/ssh_host_dsa_key" ] || ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
           eos
         when 'gentoo-paludis'
           <<-eos
             RUN cave sync
             RUN cave resolve -zx net-misc/openssh app-admin/sudo
-            RUN ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
-            RUN ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
+            RUN [ -f "/etc/ssh/ssh_host_rsa_key" ] || ssh-keygen -A -t rsa -f /etc/ssh/ssh_host_rsa_key
+            RUN [ -f "/etc/ssh/ssh_host_dsa_key" ] || ssh-keygen -A -t dsa -f /etc/ssh/ssh_host_dsa_key
           eos
         else
           raise ActionFailed,
@@ -223,19 +251,15 @@ module Kitchen
         end
 
         username = config[:username]
-        password = config[:password]
         public_key = IO.read(config[:public_key]).strip
         homedir = username == 'root' ? '/root' : "/home/#{username}"
 
         base = <<-eos
           RUN if ! getent passwd #{username}; then \
-                useradd -d #{homedir} -m -s /bin/bash #{username}; \
+                useradd -d #{homedir} -m -s /bin/bash -p '*' #{username}; \
               fi
-          RUN echo #{username}:#{password} | chpasswd
-          RUN echo '#{username} ALL=(ALL) NOPASSWD:ALL' >> /etc/sudoers
-          RUN mkdir -p /etc/sudoers.d
-          RUN echo '#{username} ALL=(ALL) NOPASSWD:ALL' >> /etc/sudoers.d/#{username}
-          RUN chmod 0440 /etc/sudoers.d/#{username}
+          RUN echo "#{username} ALL=(ALL) NOPASSWD: ALL" >> /etc/sudoers
+          RUN echo "Defaults !requiretty" >> /etc/sudoers
           RUN mkdir -p #{homedir}/.ssh
           RUN chown -R #{username} #{homedir}/.ssh
           RUN chmod 0700 #{homedir}/.ssh
@@ -247,7 +271,7 @@ module Kitchen
         Array(config[:provision_command]).each do |cmd|
           custom << "RUN #{cmd}\n"
         end
-        ssh_key = "RUN echo '#{public_key}' >> #{homedir}/.ssh/authorized_keys"
+        ssh_key = "RUN echo #{Shellwords.escape(public_key)} >> #{homedir}/.ssh/authorized_keys"
         # Empty string to ensure the file ends with a newline.
         [from, env_variables, platform, base, custom, ssh_key, ''].join("\n")
       end
@@ -269,19 +293,21 @@ module Kitchen
           end
         end
         raise ActionFailed,
-        'Could not parse Docker build output for image ID'
+          'Could not parse Docker build output for image ID'
       end
 
       def build_image(state)
         cmd = "build"
         cmd << " --no-cache" unless config[:use_cache]
+        extra_build_options = config_to_options(config[:build_options])
+        cmd << " #{extra_build_options}" unless extra_build_options.empty?
         dockerfile_contents = dockerfile
         build_context = config[:build_context] ? '.' : '-'
         file = Tempfile.new('Dockerfile-kitchen', Dir.pwd)
         output = begin
           file.write(dockerfile)
           file.close
-          docker_command("#{cmd} -f #{file.path} #{build_context}", :input => dockerfile_contents)
+          docker_command("#{cmd} -f #{Shellwords.escape(dockerfile_path(file))} #{build_context}", :input => dockerfile_contents)
         ensure
           file.close unless file.closed?
           file.unlink
@@ -307,7 +333,6 @@ module Kitchen
         Array(config[:volumes_from]).each {|container| cmd << " --volumes-from #{container}"}
         Array(config[:links]).each {|link| cmd << " --link #{link}"}
         Array(config[:devices]).each {|device| cmd << " --device #{device}"}
-        Array(config[:tmpfs]).each {|tmpfs| cmd << " --tmpfs #{tmpfs}"}
         cmd << " --name #{config[:instance_name]}" if config[:instance_name]
         cmd << " -P" if config[:publish_all]
         cmd << " -h #{config[:hostname]}" if config[:hostname]
@@ -319,6 +344,8 @@ module Kitchen
         Array(config[:cap_add]).each {|cap| cmd << " --cap-add=#{cap}"} if config[:cap_add]
         Array(config[:cap_drop]).each {|cap| cmd << " --cap-drop=#{cap}"} if config[:cap_drop]
         Array(config[:security_opt]).each {|opt| cmd << " --security-opt=#{opt}"} if config[:security_opt]
+        extra_run_options = config_to_options(config[:run_options])
+        cmd << " #{extra_run_options}" unless extra_run_options.empty?
         cmd << " #{image_id} #{config[:run_command]}"
         cmd
       end
@@ -335,11 +362,11 @@ module Kitchen
 
       def parse_container_ssh_port(output)
         begin
-          host, port = output.split(':')
+          _host, port = output.split(':')
           port.to_i
         rescue
           raise ActionFailed,
-          'Could not parse Docker port output for container SSH port'
+            'Could not parse Docker port output for container SSH port'
         end
       end
 
@@ -363,6 +390,30 @@ module Kitchen
         image_id = state[:image_id]
         docker_command("rmi #{image_id}")
       end
+
+      # Convert the config input for `:build_options` or `:run_options` in to a
+      # command line string for use with Docker.
+      #
+      # @since 2.5.0
+      # @param config [nil, String, Array, Hash] Config data to convert.
+      # @return [String]
+      def config_to_options(config)
+        case config
+        when nil
+          ''
+        when String
+          config
+        when Array
+          config.map {|c| config_to_options(c) }.join(' ')
+        when Hash
+          config.map {|k, v| Array(v).map {|c| "--#{k}=#{Shellwords.escape(c)}" }.join(' ') }.join(' ')
+        end
+      end
+
+      def dockerfile_path(file)
+        config[:build_context] ? Pathname.new(file.path).relative_path_from(Pathname.pwd).to_s : file.path
+      end
+
     end
   end
 end