+++ test/integration/inspec/inspec_spec.rb
@@ -14,10 +14,7 @@
 # limitations under the License.
 #
 
-require 'serverspec'
-set :backend, :exec
-
-describe command('/sbin/ifconfig eth0 multicast') do
-  its(:exit_status) { is_expected.to_not eq 0 }
-  its(:stderr) { is_expected.to match /Operation not permitted/ }
+# Just make sure the image launched and is reachable.
+describe command('true') do
+  its(:exit_status) { is_expected.to eq 0 }
 end